[feature] Can now opt out of mandatory 80→443 redirection
Træfik has the capability to “sniff” incoming requests to determine
whether they are HTTP or HTTP/S, and therefore can serve both on the
same port. This is useful in some situations (e.g. the `/unattended/`
URLs for Foreman).

- Teach Traefik to serve both HTTP and HTTP/S traffic on the same
port (actually, un-teach it the forcing of HTTP/S on the websecure
port). This is guarded by an Ansible configuration variable that
defaults to false